+++ blog_entries/json_entries/02-22-20.json
@@ -1,4 +1,5 @@
 {
+    "filename": "02-22-20",
     "date": "2/22/20",
     "title": "solo trip to Taiwan",
     "tags": [
