+++ blog_entries/json_entries/06-04-21.json
@@ -1,4 +1,5 @@
 {
+    "filename": "06-04-21",
     "date": "6/4/21",
     "title": "trip to Lake Tahoe",
     "tags": [
