+++ blog_entries/json_entries/10-20-19.json
@@ -1,4 +1,5 @@
 {
+    "filename": "10-20-19",
     "date": "10/20/2019",
     "tags": [
         "stuff i'm working on",
