+++ blog_entries/json_entries/10-22-19.json
@@ -1,4 +1,5 @@
 {
+    "filename": "10-22-19",
     "date": "10/22/2019",
     "tags": [
         "stuff",
