+++ blog_entries/json_entries/11-16-19.json
@@ -1,4 +1,5 @@
 {
+    "filename": "11-16-19",
     "title": "some new updates",
     "date": "11/16/19",
     "tags": [
