+++ blog_entries/json_entries/11-9-19.json
@@ -1,4 +1,5 @@
 {
+    "filename": "11-9-19",
     "date": "11/9/2019",
     "tags": [
         "stuff"
