+++ blog_entries/json_entries/test2.json
@@ -1,4 +1,5 @@
 {
+	"filename": "test2",
 	"content": "<h3> hey hey this is a another test whooo!</h3> <br><br><p>there's nothing here right now unfortunately. :(</p>",
 	"date": "10/19/2019",
 	"tags": ["test", "second"]
